Refine patient tabs structure for AB#16406
