+++ Apps/Admin/Client/Components/Details/AccountTab.razor
@@ -1,0 +1,37 @@
+﻿@using HealthGateway.Admin.Common.Constants
+@using HealthGateway.Admin.Client.Components.Support
+@inherits Fluxor.Blazor.Web.Components.FluxorComponent
+
+@if (Patient != null)
+{
+    @if (Patient.Status == PatientStatus.Default)
+    {
+        <MudText Class="mb-4 mt-6" Typo="Typo.h5">
+            Patient Details
+        </MudText>
+        <MudGrid Spacing="2">
+            <MudItem xs="12" lg="6">
+                <HgField Label="Account Created" Value="@ProfileCreatedDateTime.ToString()" data-testid="profile-created-datetime" />
+            </MudItem>
+            <MudItem xs="12" lg="6">
+                <HgField Label="Last Login" Value="@ProfileLastLoginDateTime.ToString()" data-testid="profile-last-login-datetime" />
+            </MudItem>
+            @if (CanViewHdid)
+            {
+                <MudItem xs="12">
+                    <HgField Label="HDID" Value="@Patient.Hdid" data-testid="patient-hdid" />
+                </MudItem>
+            }
+        </MudGrid>
+    }
+    else
+    {
+        <MudText data-testid="no-hg-profile">
+            No Health Gateway Profile
+        </MudText>
+    }
+}
+@if (CanViewMessagingVerifications)
+{
+    <MessageVerificationTable Data="@MessagingVerifications" IsLoading="@PatientSupportDetailsLoading" />
+}
