+++ Apps/Admin/Client/Components/Details/NotesTab.razor
@@ -1,0 +1,7 @@
+﻿@using HealthGateway.Admin.Client.Components.AgentAudit
+@inherits Fluxor.Blazor.Web.Components.FluxorComponent
+
+@if (CanViewAgentAuditHistory)
+{
+    <AgentAuditHistory Title="Notes" AgentActions="@AgentAuditHistory" IsLoading="@PatientSupportDetailsLoading" />
+}
