+++ Apps/Admin/Client/Components/Support/DatasetAccessSection.razor
@@ -1,7 +1,7 @@
 ﻿@using HealthGateway.Common.Data.Constants
 @inherits Fluxor.Blazor.Web.Components.FluxorComponent
 
-<MudText Class="mt-4" Typo="Typo.subtitle1">
+<MudText Class="mb-4 mt-6" Typo="Typo.h5">
     Dataset Access
 </MudText>
