+++ Apps/Admin/Client/Components/Support/DependentTable.razor
@@ -1,7 +1,7 @@
 @using HealthGateway.Common.Data.Utils
 @inherits Fluxor.Blazor.Web.Components.FluxorComponent
 
-<MudText Class="mt-4" Typo="Typo.subtitle1">
+<MudText Class="mb-4 mt-6" Typo="Typo.h5">
     Dependents
 </MudText>
